% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptPartitioning2D.R
\name{GenData2D}
\alias{GenData2D}
\title{GenData2D}
\usage{
GenData2D(n, changepoints, means1, means2, noise1 = 1, noise2 = 1)
}
\arguments{
\item{n}{number of data point.}

\item{changepoints}{a vector of increasing changepoint indices (last element is always n!).}

\item{means1}{vector of successive means1 for data1.}

\item{means2}{vector of successive means2 for data2.}

\item{noise1}{standard deviation of an additional normal noise (for data1).}

\item{noise2}{standard deviation of an additional normal noise (for data2).}
}
\value{
matrix of data of dimension 2 x n with a given values of means by the segmentation.
}
\description{
Generation of data of dimension 2 with a given values of means and changepoints
}
\examples{
GenData2D (n = 10, changepoints = c(2, 4, 6, 8, 10), means1 = c (0, 1, 0, 1, 0), means2 = (1, 2, 3, 4, 5), noise1 = 1, noise2 = 1)
}
