% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OptPart2D}
\alias{OptPart2D}
\title{OptPart2D}
\usage{
OptPart2D(data1, data2, penalty, type)
}
\arguments{
\item{data1}{is a vector of data1(a univariate time series).}

\item{data2}{is a vector of data2(a univariate time series).}

\item{penalty}{is a value of penalty (a non-negative real number).}

\item{type}{is a string defining the  algorithm ("null" = Optimal Partitioning, "pruning" = PELT).}
}
\value{
a list of 4 elements  = (changepoints, means1, means2, GlobalCost).                    
 
\describe{
\item{\code{changepoints}}{is the vector of changepoints.}
\item{\code{means1}}{is the vector of successive means for data1.}
\item{\code{means2}}{is the vector of successive means for data2.}
\item{\code{globalCost}}{is a number equal to the global cost.}
}                                                                                                                                                                             #
}
\description{
Optimal Partitioning and PELT algorithms for changepoints dimension 2
}
\examples{
OptPart2D(data1 = c(0,0,0,1,1,1), data2 = c(2,2,2,0,0,0), penalty = 2*log(6),  type = "null") 
}
